MLDBFB-615: split event loop and tcp bits from http into libio
