+++ http/testing/http_service_bench.cc
@@ -17,11 +17,11 @@
 #include <boost/program_options/parsers.hpp>
 #include <boost/program_options/variables_map.hpp>
 
-#include "mldb/http/asio_thread_pool.h"
-#include "mldb/http/event_loop.h"
+#include "mldb/io/asio_thread_pool.h"
+#include "mldb/io/event_loop.h"
 #include "mldb/http/http_socket_handler.h"
-#include "mldb/http/port_range_service.h"
-#include "mldb/http/tcp_acceptor.h"
+#include "mldb/io/port_range_service.h"
+#include "mldb/io/tcp_acceptor.h"
 
 
 using namespace std;
