+++ io/tcp_acceptor.cc
@@ -10,7 +10,7 @@
 
 #include "mldb/arch/exception.h"
 #include "mldb/base/exc_assert.h"
-#include "mldb/http/tcp_acceptor_impl.h"
+#include "mldb/io/tcp_acceptor_impl.h"
 #include "event_loop.h"
 #include "port_range_service.h"
 #include "tcp_socket_handler.h"
