+++ io/tcp_acceptor_impl.cc
@@ -14,10 +14,10 @@
 #include "mldb/base/exc_assert.h"
 #include "mldb/types/url.h"
 #include "mldb/base/scope.h"
-#include "mldb/http/port_range_service.h"
-#include "mldb/http/event_loop.h"
-#include "mldb/http/tcp_acceptor.h"
-#include "mldb/http/tcp_socket.h"
+#include "mldb/io/port_range_service.h"
+#include "mldb/io/event_loop.h"
+#include "mldb/io/tcp_acceptor.h"
+#include "mldb/io/tcp_socket.h"
 #include "event_loop_impl.h"
 #include "tcp_socket_impl.h"
 #include "tcp_acceptor_impl.h"
