+++ rest/asio_peer_server.h
@@ -9,7 +9,7 @@
 #pragma once
 
 #include "peer_connection.h"
-#include "mldb/http/port_range_service.h"
+#include "mldb/io/port_range_service.h"
 
 namespace Datacratic {
