+++ rest/http_rest_service.h
@@ -12,9 +12,9 @@
 #include <memory>
 #include <string>
 
-#include "mldb/http/asio_thread_pool.h"
+#include "mldb/io/asio_thread_pool.h"
 #include "mldb/rest/http_rest_endpoint.h"
-#include "mldb/http/port_range_service.h"
+#include "mldb/io/port_range_service.h"
 #include "mldb/rest/rest_connection.h"
 #include "mldb/rest/rest_request.h"
 #include "mldb/types/date.h"
