+++ soa/service/testing/asio_timer_test.cc
@@ -7,9 +7,9 @@
 
 */
 
-#include "mldb/http/event_loop.h"
-#include "mldb/http/asio_timer.h"
-#include "mldb/http/asio_thread_pool.h"
+#include "mldb/io/event_loop.h"
+#include "mldb/io/asio_timer.h"
+#include "mldb/io/asio_thread_pool.h"
 #include "mldb/watch/watch_impl.h"
 #include <boost/asio.hpp>
 #include <thread>
